Also describe editor tools in disabled_tool_info when disabled
