+++ .ignore.eslint
@@ -1611,7 +1611,6 @@ packages/lib/services/ai/SearchService.js
 packages/lib/services/ai/SqliteVec.test.js
 packages/lib/services/ai/aiSettingsTransition.test.js
 packages/lib/services/ai/aiSettingsTransition.js
-packages/lib/services/ai/applyNoteEdits.js
 packages/lib/services/ai/availability.test.js
 packages/lib/services/ai/availability.js
 packages/lib/services/ai/chunker.test.js
@@ -1629,18 +1628,24 @@ packages/lib/services/ai/providers/OpenAiCompatible.js
 packages/lib/services/ai/providers/TestProvider.js
 packages/lib/services/ai/testing/TestEmbeddingProvider.test.js
 packages/lib/services/ai/testing/TestEmbeddingProvider.js
-packages/lib/services/ai/tools/createNote.js
-packages/lib/services/ai/tools/createNotebook.js
-packages/lib/services/ai/tools/deleteNote.js
-packages/lib/services/ai/tools/index.js
-packages/lib/services/ai/tools/listNotebooks.js
-packages/lib/services/ai/tools/listTags.js
-packages/lib/services/ai/tools/manageTags.js
-packages/lib/services/ai/tools/readNote.js
-packages/lib/services/ai/tools/searchNotes.js
-packages/lib/services/ai/tools/semanticSearchNotes.js
-packages/lib/services/ai/tools/updateNote.js
+packages/lib/services/ai/tools/ToolIndex.js
+packages/lib/services/ai/tools/buildEditorTools.js
+packages/lib/services/ai/tools/global/createNote.js
+packages/lib/services/ai/tools/global/createNotebook.js
+packages/lib/services/ai/tools/global/deleteNote.js
+packages/lib/services/ai/tools/global/index.js
+packages/lib/services/ai/tools/global/listNotebooks.js
+packages/lib/services/ai/tools/global/listTags.js
+packages/lib/services/ai/tools/global/manageTags.js
+packages/lib/services/ai/tools/global/readNote.js
+packages/lib/services/ai/tools/global/searchNotes.js
+packages/lib/services/ai/tools/global/semanticSearchNotes.js
+packages/lib/services/ai/tools/global/updateNote.js
+packages/lib/services/ai/tools/types.js
+packages/lib/services/ai/tools/utils/applyNoteEdits.test.js
+packages/lib/services/ai/tools/utils/applyNoteEdits.js
 packages/lib/services/ai/tools/utils/buildTool.js
+packages/lib/services/ai/tools/utils/settings.js
 packages/lib/services/ai/types.js
 packages/lib/services/ai/utils/findFencedBlock.js
 packages/lib/services/commands/MenuUtils.js
