+++ packages/lib/services/ai/tools/global/index.ts
@@ -1,0 +1,23 @@
+import searchNotes from './searchNotes';
+import semanticSearchNotes from './semanticSearchNotes';
+import readNote from './readNote';
+import listNotebooks from './listNotebooks';
+import listTags from './listTags';
+import createNote from './createNote';
+import updateNote from './updateNote';
+import deleteNote from './deleteNote';
+import manageTags from './manageTags';
+import createNotebook from './createNotebook';
+
+export default [
+	searchNotes,
+	semanticSearchNotes,
+	readNote,
+	listNotebooks,
+	listTags,
+	createNote,
+	updateNote,
+	deleteNote,
+	manageTags,
+	createNotebook,
+];
